requirements.txt
notebook==6.5.4
pre-commit==2.13.0
boto3==1.18.30
numpy
scipy
openpyxl
xlrd
xlwt
matplotlib
pandas
Pillow
opencv-python
PyPDF2
pdfminer.six
sympy
scikit-learn
flask
gunicorn
requests
gevent
flask-cors
python-docx
pdfplumber
loguru
redis-py-cluster
dill